+++ src/budy/models/measurement.py
@@ -244,6 +244,7 @@ class Measurement(base.BudyBase):
 
         meta = dict(
             object_id = object_id,
+            company_product_code = company_product_code,
             modify_date = modify_date,
             discount = discount
         )