+++ CHANGELOG.md
@@ -9,9 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* Support for the `invoiced` state at the `Order` model
-* Storage of `company_product_code` inside `Product` and `Measurement` on import data
-* Support for price fixing on Omni import
+*
 
 ### Changed
 
@@ -20,6 +18,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 *
+
+## [0.5.0] - 2021-03-14
+
+### Added
+
+* Support for the `invoiced` state at the `Order` model
+* Storage of `company_product_code` inside `Product` and `Measurement` on import data
+* Support for price fixing on Omni import
 
 ## [0.4.2] - 2021-02-25